ml_models/model_ner/requirements.txt
flask==2.2.5
flask-cors==3.0.10
torch==2.0.0
transformers==4.30.0
azure-storage-blob
requests==2.28.1
pydantic==1.10.5
python-decouple
aimped
numpy==1.24.4
nltk==3.7
ipython
seqeval
flasgger[all]
sentencepiece
protobuf==3.20.3